Added holdButtonDisabled function to SignInForm
